+++ brownie_tokens/forked.py
@@ -1,12 +1,29 @@
+import os
 import requests
 import sys
 from brownie import Contract, Wei
 from brownie.convert import to_address
-from typing import Dict, List
+from typing import Dict, List, Optional
+
+_ethplorer_api_key: str = os.getenv("ETHPLORER_API_KEY") or "freekey"
 
 _token_holders: Dict = {}
+_skip_list: List[str] = []
 
 _token_names = ["Aave"]
+
+
+class BrownieTokensError(Exception):
+    pass
+
+
+def update_skipped_addresses(token: Optional[str] = None) -> None:
+    if token:
+        for address in _skip_list:
+            _token_holders[token].remove(address)
+    else:
+        for token in _token_holders:
+            update_skipped_addresses(token)
 
 
 def get_top_holders(address: str) -> List:
@@ -14,14 +31,36 @@ def get_top_holders(address: str) -> List:
     if address not in _token_holders:
         holders = requests.get(
             f"https://api.ethplorer.io/getTopTokenHolders/{address}",
-            params={"apiKey": "freekey", "limit": "50"},
+            params={"apiKey": _ethplorer_api_key, "limit": "50"},
         ).json()
+        if "error" in holders:
+            api_key_message = ""
+            if _ethplorer_api_key == "freekey":
+                api_key_message = (
+                    " Adding $ETHPLORER_API_KEY (from https://ethplorer.io) as environment variable"
+                    " may solve the problem."
+                )
+            raise BrownieTokensError(
+                f"Ethplorer returned error: {holders['error']}." + api_key_message
+            )
+
         _token_holders[address] = [to_address(i["address"]) for i in holders["holders"]]
         if address in _token_holders[address]:
             # don't steal from the treasury - that could cause wierdness
             _token_holders[address].remove(address)
+        update_skipped_addresses(address)
 
     return _token_holders[address]
+
+
+def skip_holders(*addresses: str) -> None:
+    """
+    Addresses that will remain untouched via '_mint_for_testing'.
+    Does not include minting with custom logic.
+    """
+    global _skip_list
+    _skip_list = list(set(_skip_list + list(addresses)))
+    update_skipped_addresses()
 
 
 class MintableForkToken(Contract):